add validation and ExceptionTranslator

+++ User/src/main/java/com/example/User/Service/UserService.java
@@ -4,10 +4,14 @@ import com.example.User.Entity.Users;
 import com.example.User.Repository.UserRepository;
 import com.example.User.VO.Department;
 import com.example.User.VO.ResponseTemplateVO;
+import com.example.User.models.UserDto;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.util.StringUtils;
 import org.springframework.web.client.RestTemplate;
+
+import java.util.Optional;
 
 @Service
 @Slf4j
@@ -18,14 +22,36 @@ public class UserService {
     @Autowired
     private RestTemplate restTemplate;
 
-    public Users saveUser(Users users) {
+    public Users saveUser(UserDto userDto) {
+
+        validate(userDto);
+        Users users = new Users();
+        users.setName(userDto.getName());
+        users.setEmail(userDto.getEmail());
+        users.setDepartmentId(userDto.getDepartmentId());
 
         log.info("Inside saveUser method of UserService");
         return userRepository.save(users);
 
     }
 
+    private void validate(UserDto userDto) {
+        if (StringUtils.isEmpty(userDto.getName())) {
+            throw new RuntimeException("name is empty!");
+        }
+
+
+    }
+
     public void deleteUserById(Long userId) {
+
+        if (userId == null) {
+            throw new RuntimeException("userId is null!");
+        }
+        Optional<Users> byId = userRepository.findById(userId);
+        if (byId.isEmpty()) {
+            throw new RuntimeException("user not found!");
+        }
 
         log.info("Inside deleteUserById method of UserService");
         userRepository.deleteById(userId);
@@ -50,7 +76,7 @@ public class UserService {
         return vo;
     }
 
-    public ResponseTemplateVO getUserWIthDepartment(){
+    public ResponseTemplateVO getUserWIthDepartment() {
         ResponseTemplateVO vo = new ResponseTemplateVO();
         log.info("Inside getUserWithDepartment method of UserService");
         return vo;
